Document RBF ChgDir, and unify PEXEC mode handling in ChgDir.
Mostly this is to add more documentation about how ChgDir
works on RBF devices, in preparation for writing other
File Managers, especially FUSE or remote file managers,
that do not use the RBF disk format.

I made one small change to Level2:
In Level1, you can change the current execution directory (CXD)
by setting either the EXEC or PEXEC mode on the
I$ChgDir system call.  In Level2, it was only
accepting EXEC.  This commit changes Level2 so that
it also accepts PEXEC.

I think this is really minor, and that PEXEC was probably
never documented or used in the I$ChgDir API.
But unifying Levels 1 and 2 is a goal,
and I don't want to break any existing code.