Setup clean and smudge filters
Commit clean and smudge filters into the repo and the
setup-filters script is added.  The script can be used to
setup the filters and must be run from a shell command
line in the nitros9 repo base directory.  To run the setup
type "scripts/setup-filters"

Filters require python3 in PATH and the script requires /bin/sh

On initial commit the script clean filter defaults to
the "spread out" source formatting currently in github.

After the github repository is converted to "condensed"
formatting the local repo must also be converted and the
setup-filters line "#ASMCLEAN=$REPODIR/scripts/asmclean"
should be uncommented.