Param to allow Pausing to default off on UARTS.
With modern terminal emulators, the need for automatic pausing
is less than its annoyance factor.  It also interferes
with automatic testing and benchmarking.

This does not change the default value,
but lets you change it with -D from your makefile.

This is for the /TERM device with three UART drivers,
mc6850, sc6850, and sc6551.